requirements/dev.txt
-r base.txt

black==23.12.1
isort==5.13.2
flake8==6.1.0
django-extensions==3.2.3
